Improve overlay detection by looking for > 50% word match between rom name and overlay name.
Only case that is not working is "SpinBall" vs "Spin Ball" mismatch, but this is hard to detect as it's just plainly the wrong name in the overlay name.

+++ src/Overlays.h
@@ -14,7 +14,7 @@ public:
 
 private:
     // Returns a confidence level; 0 means no match, 1..n is increasing confidence score
-    int OverlayFileFuzzyMatch(const fs::path& p1, const fs::path& p2);
+    float OverlayFileFuzzyMatch(const fs::path& p1, const fs::path& p2);
 
     struct OverlayFile {
         fs::path path;
